+++ client/src/backend/remote_node.rs
@@ -197,6 +197,18 @@ impl backend::Backend for RemoteNode {
         Ok(keys.into_iter().map(|key| key.0).collect())
     }
 
+    async fn block_header(&self, block_hash: Option<BlockHash>) -> Result<BlockHeader, Error> {
+        self.rpc
+            .chain
+            .header(block_hash)
+            .compat()
+            .await?
+            .ok_or_else(|| match block_hash {
+                Some(hash) => format!("Header not found for block hash {}", hash).into(),
+                None => "Header not found for the best chain tip".into(),
+            })
+    }
+
     fn get_genesis_hash(&self) -> Hash {
         self.genesis_hash
     }